Split apart draft note publisher + manager
