+++ cmd/app/draft_notes.go
@@ -51,8 +51,20 @@ func (draftNote DraftNoteWithPosition) GetPositionData() PositionData {
 	return draftNote.PositionData
 }
 
+type DraftNoteManager interface {
+	ListDraftNotes(pid interface{}, mergeRequest int, opt *gitlab.ListDraftNotesOptions, options ...gitlab.RequestOptionFunc) ([]*gitlab.DraftNote, *gitlab.Response, error)
+	CreateDraftNote(pid interface{}, mergeRequest int, opt *gitlab.CreateDraftNoteOptions, options ...gitlab.RequestOptionFunc) (*gitlab.DraftNote, *gitlab.Response, error)
+	DeleteDraftNote(pid interface{}, mergeRequest int, note int, options ...gitlab.RequestOptionFunc) (*gitlab.Response, error)
+	UpdateDraftNote(pid interface{}, mergeRequest int, note int, opt *gitlab.UpdateDraftNoteOptions, options ...gitlab.RequestOptionFunc) (*gitlab.DraftNote, *gitlab.Response, error)
+}
+
+type draftNoteService struct {
+	data
+	client DraftNoteManager
+}
+
 /* draftNoteHandler creates, edits, and deletes draft notes */
-func (a *Api) draftNoteHandler(w http.ResponseWriter, r *http.Request) {
+func (a draftNoteService) handler(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 	switch r.Method {
 	case http.MethodGet:
@@ -69,64 +81,29 @@ func (a *Api) draftNoteHandler(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-type DraftNotePublisher interface {
-	PublishAllDraftNotes(pid interface{}, mergeRequest int, options ...gitlab.RequestOptionFunc) (*gitlab.Response, error)
-	PublishDraftNote(pid interface{}, mergeRequest int, note int, options ...gitlab.RequestOptionFunc) (*gitlab.Response, error)
-}
+/* listDraftNotes lists all draft notes for the currently authenticated user */
+func (a draftNoteService) listDraftNotes(w http.ResponseWriter, _ *http.Request) {
 
-type draftNoteService struct {
-	data
-	client DraftNotePublisher
-}
-
-func (a draftNoteService) handler(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	if r.Method != http.MethodPost {
-		w.Header().Set("Access-Control-Allow-Methods", http.MethodPost)
-		handleError(w, InvalidRequestError{}, "Expected POST", http.StatusMethodNotAllowed)
-		return
-	}
-
-	body, err := io.ReadAll(r.Body)
-	if err != nil {
-		handleError(w, err, "Could not read request body", http.StatusBadRequest)
-		return
-	}
-
-	defer r.Body.Close()
-	var draftNotePublishRequest DraftNotePublishRequest
-	err = json.Unmarshal(body, &draftNotePublishRequest)
+	opt := gitlab.ListDraftNotesOptions{}
+	draftNotes, res, err := a.client.ListDraftNotes(a.projectInfo.ProjectId, a.projectInfo.MergeId, &opt)
 
 	if err != nil {
-		handleError(w, err, "Could not read JSON from request", http.StatusBadRequest)
-		return
-	}
-
-	var res *gitlab.Response
-	if draftNotePublishRequest.PublishAll {
-		res, err = a.client.PublishAllDraftNotes(a.projectInfo.ProjectId, a.projectInfo.MergeId)
-	} else {
-		if draftNotePublishRequest.Note == 0 {
-			handleError(w, errors.New("No ID provided"), "Must provide Note ID", http.StatusBadRequest)
-			return
-		}
-		res, err = a.client.PublishDraftNote(a.projectInfo.ProjectId, a.projectInfo.MergeId, draftNotePublishRequest.Note)
-	}
-
-	if err != nil {
-		handleError(w, err, "Could not publish draft note(s)", http.StatusInternalServerError)
+		handleError(w, err, "Could not get draft notes", http.StatusInternalServerError)
 		return
 	}
 
 	if res.StatusCode >= 300 {
-		handleError(w, GenericError{endpoint: "/mr/draft_notes/publish"}, "Could not publish dfaft note", res.StatusCode)
+		handleError(w, GenericError{endpoint: "/mr/draft/comment"}, "Could not get draft notes", res.StatusCode)
 		return
 	}
 
 	w.WriteHeader(http.StatusOK)
-	response := SuccessResponse{
-		Message: "Draft note(s) published",
-		Status:  http.StatusOK,
+	response := ListDraftNotesResponse{
+		SuccessResponse: SuccessResponse{
+			Message: "Draft notes fetched successfully",
+			Status:  http.StatusOK,
+		},
+		DraftNotes: draftNotes,
 	}
 
 	err = json.NewEncoder(w).Encode(response)
@@ -136,7 +113,7 @@ func (a draftNoteService) handler(w http.ResponseWriter, r *http.Request) {
 }
 
 /* postDraftNote creates a draft note */
-func (a *Api) postDraftNote(w http.ResponseWriter, r *http.Request) {
+func (a draftNoteService) postDraftNote(w http.ResponseWriter, r *http.Request) {
 	body, err := io.ReadAll(r.Body)
 	if err != nil {
 		handleError(w, err, "Could not read request body", http.StatusBadRequest)
@@ -194,7 +171,7 @@ func (a *Api) postDraftNote(w http.ResponseWriter, r *http.Request) {
 }
 
 /* deleteDraftNote deletes a draft note */
-func (a *Api) deleteDraftNote(w http.ResponseWriter, r *http.Request) {
+func (a draftNoteService) deleteDraftNote(w http.ResponseWriter, r *http.Request) {
 	suffix := strings.TrimPrefix(r.URL.Path, "/mr/draft_notes/")
 	id, err := strconv.Atoi(suffix)
 	if err != nil {
@@ -227,7 +204,7 @@ func (a *Api) deleteDraftNote(w http.ResponseWriter, r *http.Request) {
 }
 
 /* updateDraftNote edits the text of a draft comment */
-func (a *Api) updateDraftNote(w http.ResponseWriter, r *http.Request) {
+func (a draftNoteService) updateDraftNote(w http.ResponseWriter, r *http.Request) {
 	suffix := strings.TrimPrefix(r.URL.Path, "/mr/draft_notes/")
 	id, err := strconv.Atoi(suffix)
 	if err != nil {
@@ -279,37 +256,6 @@ func (a *Api) updateDraftNote(w http.ResponseWriter, r *http.Request) {
 			Status:  http.StatusOK,
 		},
 		DraftNote: draftNote,
-	}
-
-	err = json.NewEncoder(w).Encode(response)
-	if err != nil {
-		handleError(w, err, "Could not encode response", http.StatusInternalServerError)
-	}
-}
-
-/* listDraftNotes lists all draft notes for the currently authenticated user */
-func (a *Api) listDraftNotes(w http.ResponseWriter, _ *http.Request) {
-
-	opt := gitlab.ListDraftNotesOptions{}
-	draftNotes, res, err := a.client.ListDraftNotes(a.projectInfo.ProjectId, a.projectInfo.MergeId, &opt)
-
-	if err != nil {
-		handleError(w, err, "Could not get draft notes", http.StatusInternalServerError)
-		return
-	}
-
-	if res.StatusCode >= 300 {
-		handleError(w, GenericError{endpoint: "/mr/draft/comment"}, "Could not get draft notes", res.StatusCode)
-		return
-	}
-
-	w.WriteHeader(http.StatusOK)
-	response := ListDraftNotesResponse{
-		SuccessResponse: SuccessResponse{
-			Message: "Draft notes fetched successfully",
-			Status:  http.StatusOK,
-		},
-		DraftNotes: draftNotes,
 	}
 
 	err = json.NewEncoder(w).Encode(response)
